requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.59
botocore==1.35.59
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.16
django-allauth==64.2.0
django-autocomplete-light==3.11.0
django-extensions==3.2.3
django-otp==1.5.4
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
kombu==5.4.2
python-dotenv==1.0.1
redis==5.2.0
s3transfer==0.10.3
sqlparse==0.5.1
stripe==11.2.0
tzdata==2024.2
vine==5.1.0
whitenoise==6.8.2
psycopg2-binary==2.9.9
